Distributor.py: Added self.packages attribute for persistent tracking of distributed packages. Initialized with all package.pid values, each value is removed from the attribute as it is loaded into a truck. This removes the possibility of duplicate deliveries. Duplication has been a problem.
package_handler.py: Added logging for debugging.

interface.py: Now self-adjusts the frame and the columns based on the content.

time_sim.py: Employed deepcopy() for copying states to the minute objects, which resolved the issue of the interface only displaying the final state of the package and truck objects. The copy method previously used was not making new copies of the objects and thereby not saving snapshots of their states.